allow deleting both digits in minutes and hours entry
Before this commit, users couldn't delete both digits; for example, to change 0 to 45, they had to first type 04, then delete the 0 and only then could they add the 5 (since each input is limited to two digits and can't be empty).

Now the user can delete both digits, and new functions were added to type a zero if the field is deselected and empty, or if enter is pressed when the field is empty.